third_party/blink/web_tests/external/wpt/css/css-mixins/contents-rule-expected.txt

This is a testharness.js-based test.
[FAIL] @contents is ignored if there is no @contents parameter
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
